added apology.html and update.html

+++ app.py
@@ -6,7 +6,7 @@ from tempfile import mkdtemp
 from sqlalchemy.sql.expression import select
 from werkzeug.security import check_password_hash, generate_password_hash
 from datetime import datetime
-from helpers import login_required
+from helpers import login_required ,apology
 
 
 # Configure application
@@ -39,10 +39,6 @@ def index():
         print(task)
         return render_template("index.html", task=task)
       
-
-
-    
-
 @app.route("/contact", methods=["GET", "POST"])
 def contact():
     if request.method == "POST":
@@ -56,10 +52,14 @@ def login():
     """Log user in"""
     session.clear()
     if request.method == "POST":
+        # Ensure username was submitted
         if not request.form.get("username"):
-            return "must provide username"
+            return apology("must provide username", 403)
+
+        # Ensure password was submitted
         elif not request.form.get("password"):
-            return "must provide password"
+            return apology("must provide password", 403)
+
         rows = db.execute("SELECT * FROM users WHERE username = ?", request.form.get("username"))
 
         # Ensure username exists and password is correct
@@ -98,13 +98,19 @@ def register():
         c_password=request.form.get("confirmation")
         #show apology if some error was occured
         if not username:
-            return "must provide username"
+            return apology("must provide username",400)
         elif not password or not  c_password :
-            return"must provide password" 
+            return apology("must provide password" ,400)
+        #implemented the regex's function with the help  of stackoverflow
+
+        #MAKE SURE BOTH PASSWORD MATCH
+        elif  password !=  c_password:
+            return apology("both password  must match", 400)
+
     
         rows = db.execute("SELECT * FROM users WHERE username = :username", username=request.form.get("username"))
         if len(rows) >= 1:
-            return ("username already exists" , 400)
+            return apology("username already exists" , 400)
         # Start session
         db.execute("INSERT INTO users (username, hash) VALUES (:username, :hash)",username=request.form.get("username"),
                              hash=generate_password_hash(request.form.get("password")))
@@ -123,10 +129,20 @@ def register():
 def delete():
     if request.method == "POST":
         user_id=session["user_id"]
+       # Ensure username was submitted
         if not request.form.get("username"):
-            return "must provide username"
-        elif not  request.form.get("password"):
-            return"must provide password" 
+            return apology("must provide username", 403)
+
+        # Ensure password was submitted
+        elif not request.form.get("password"):
+            return apology("must provide password", 403)
+        # Query database for username
+        rows = db.execute("SELECT * FROM users WHERE username = ?", request.form.get("username"))
+
+        # Ensure username exists and password is correct
+        if len(rows) != 1 or not check_password_hash(rows[0]["hash"], request.form.get("password")):
+            return apology("invalid username and/or password", 403)
+
         db.execute('DELETE FROM users WHERE id = ?', (user_id))
         session.clear()
         return redirect("/")
@@ -153,9 +169,24 @@ def  task():
     else:
         return render_template("task.html")
 
-        
-@app.route('/delete/<string:task>', methods = ['GET'])
+@app.route('/update/<string:task_id>', methods=['POST','GET'])
+def update(task_id):
+    task= db.execute("select * from task where task_id=%s" , (str(task_id)))
+    if request.method == 'POST':
+        subject = request.form.get("sub")
+        day = request.form.get("day")
+        month = request.form.get("month")
+        task = request.form.get("task")
+        date = request.form.get("date")
+        priority = request.form.get("priority")
+        db.execute("UPDATE task SET subject = %s, day= %s,month =%s, task= %s , date=%s , priority=%s WHERE task_id = %s", subject,day,month,task,date,priority,str(task_id))
+        return redirect(url_for("index"))
 
+    else:
+         return render_template("update.html", task=task)
+        
+
+@app.route('/delete/<string:task>', methods = ['GET'])
 def delete_task(task):
     db.execute("DELETE FROM task WHERE task=%s", (task))
     return redirect(url_for("index"))
